// Assets-Uncapped/front-portal-develop/src/services/api/organisation-users/models/OutstandingDebtDocumentResponse.ts
export interface OutstandingDebtDocumentResponse {
    /**
     * 
     * @type {string}
     * @memberof OutstandingDebtDocumentResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof OutstandingDebtDocumentResponse
     */
    loanProvider?: string;
    /**
     * 
     * @type {string}
     * @memberof OutstandingDebtDocumentResponse
     */
    currency?: string;
    /**
     * 
     * @type {number}
     * @memberof OutstandingDebtDocumentResponse
     */
    remainingBalance?: number;
    /**
     * 
     * @type {string}
     * @memberof OutstandingDebtDocumentResponse
     */
    repaymentType?: OutstandingDebtDocumentResponseRepaymentTypeEnum;
    /**
     * 
     * @type {Date}
     * @memberof OutstandingDebtDocumentResponse
     */
    finalRepayment?: Date;
    /**
     * 
     * @type {string}
     * @memberof OutstandingDebtDocumentResponse
     */
    securityType?: OutstandingDebtDocumentResponseSecurityTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof OutstandingDebtDocumentResponse
     */
    revenueShare?: number;
    /**
     * 
     * @type {number}
     * @memberof OutstandingDebtDocumentResponse
     */
    maximumLimit?: number;
    /**
     * 
     * @type {string}
     * @memberof OutstandingDebtDocumentResponse
     */
    status?: OutstandingDebtDocumentResponseStatusEnum;
    /**
     * 
     * @type {Date}
     * @memberof OutstandingDebtDocumentResponse
     */
    lastModificationDate?: Date;
}


/**
 * @export
 */
export const OutstandingDebtDocumentResponseRepaymentTypeEnum = {
    Amortising: 'AMORTISING',
    Bullet: 'BULLET',
    RevolvingLineOfCredit: 'REVOLVING_LINE_OF_CREDIT',
    ConvertibleLoan: 'CONVERTIBLE_LOAN',
    RevenueBasedFinancingMca: 'REVENUE_BASED_FINANCING_MCA',
    CreditCardDebt: 'CREDIT_CARD_DEBT',
    Leases: 'LEASES'
} as const;
export type OutstandingDebtDocumentResponseRepaymentTypeEnum = typeof OutstandingDebtDocumentResponseRepaymentTypeEnum[keyof typeof OutstandingDebtDocumentResponseRepaymentTypeEnum];

/**
 * @export
 */
export const OutstandingDebtDocumentResponseSecurityTypeEnum = {
    Secured: 'SECURED',
    Unsecured: 'UNSECURED'
} as const;
export type OutstandingDebtDocumentResponseSecurityTypeEnum = typeof OutstandingDebtDocumentResponseSecurityTypeEnum[keyof typeof OutstandingDebtDocumentResponseSecurityTypeEnum];

/**
 * @export
 */
export const OutstandingDebtDocumentResponseStatusEnum = {
    Uploaded: 'UPLOADED',
    Locked: 'LOCKED',
    Deleted: 'DELETED',
    Requested: 'REQUESTED',
    DeletedAutomaticDocument: 'DELETED_AUTOMATIC_DOCUMENT'
} as const;
export type OutstandingDebtDocumentResponseStatusEnum = typeof OutstandingDebtDocumentResponseStatusEnum[keyof typeof OutstandingDebtDocumentResponseStatusEnum];